Reduce factory redundancy.
